core-nonebot/requirements.txt
nonebot2==2.0.0a13.post1
python_dateutil==2.8.1
rich==10.1.0
tinydb==4.4.0
nonebot-adapter-cqhttp==2.0.0a13
nonebot-plugin-apscheduler==0.1.2
httpx~=0.17.1
python-dateutil~=2.8.1
PyMySQL==0.9.3
aiomysql==0.0.21
fastapi~=0.63.0
pydantic~=1.8.2